Renaming and linting
